Added movie projections.
New entity for movie projections called Projection. All crud operations for that entity are available only for administrators. The projection hold information about the cinema where is going to be performed, movie, date of projection (validated with custom date validator - only dates after tommorow :+1: ), the hall of the cinema and the hours of projection intself.