Non deleteing entities
For the two entities which can be deleted from the application the services no longer delete them from the data base, they just set the new property deletedOn to a new date and the repositories just checking if the value is null and return filtered result for entities which are no deleted. That affects movie and projection. If movie is deleted then the projection is no longer available. if user has reservation to the current deleted proejction - he can see it no longer.